+++ NEWS.md
@@ -4,7 +4,10 @@
 * New data for R 4.3.2
 * New column `"priority"` records whether an object's package was `"base"` or 
   `"recommended"`.
+* Column `"generic"` has been renamed to `"S4generic"`, to be more informative.
 * Some data cleanups.
+* Data is extended back to R version 0.61.
+
 
 # rcheology 4.3.1.1